requirements.txt
attrs==25.4.0
Babel==2.10.3
beautifulsoup4==4.14.3
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
cryptography==42.0.8
decorator==5.1.1
docopt==0.6.2
docutils==0.20.1
ebaysdk==2.1.5
freetype-py==2.5.1
freezegun==1.2.1
geoip2==2.9.0
gevent==24.2.1
greenlet==3.0.3
idna==3.6
isodate==0.7.2
Jinja2==3.1.2
libsass==0.22.0
lxml==5.2.1
lxml_html_clean==0.4.3
MarkupSafe==2.1.5
maxminddb==3.0.0
num2words==0.5.13
ofxparse==0.21
passlib==1.7.4
pillow==10.2.0
platformdirs==4.5.1
polib==1.1.1
psutil==5.9.8
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==2.23
pydot==1.4.2
pyOpenSSL==24.1.0
pyparsing==3.3.1
PyPDF2==2.12.1
pypng==0.20220715.0
pyserial==3.5
python-dateutil==2.8.2
python-ldap==3.4.4
python-stdnum==1.19
pytz==2025.2
pyusb==1.2.1
qrcode==7.4.2
reportlab==4.1.0
requests==2.31.0
requests-file==3.0.1
requests-toolbelt==1.0.0
rjsmin==1.2.0
rlPyCairo==0.4.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.8.1
typing_extensions==4.15.0
urllib3==2.0.7
vobject==0.9.6.1
Werkzeug==3.0.1
wheel==0.45.1
xlrd==2.0.1
XlsxWriter==3.1.9
xlwt==1.3.0
zeep==4.2.1
zope.event==6.1
zope.interface==8.1.1
